Fix getCurrentTime and seek
All the calls to seek already include audioSource.offset.
vorbisFile.timeTell() returns the time in the file which you're pulling data from, but since we're streaming, that's actually in the future. So timeBlocks are a quick and dirty way of caching what would have been the previous values of timeTell.