Avoid compiler warnings from overriding the O2 flag
Gcc and clang silently accept overrides (for `-O2 -O0`). Msvc, while accepting the override, complains about it with warning D9025 (for `/O2 /Od`).

`-O2` is included by hxcpp's various toolchains for files that have the `release` or `optim-std` tag.

For the lime/openfl haxe sources, the `haxe,release` tags are manually specified in all files unless `-D lime_debug` and `-D openfl_debug` are used. By setting those defines, and calling an initialization macro, we can set the tags for all haxe sources to `haxe,optim-none` instead.

For the C/C++ sources, however, while no optimization level is specified, there also doesn't seem to be a way to do so via command line override. When no optimization level is set, it defaults to `optim-std`.

So, the only way to completely avoid the warning is to only add our own optimization flags for files that are being compiled with the `optim-none` tag. This leaves out all C/C++ sources from quicker compilation.

Of course this is only necessary for msvc, but the change is applied to gcc as well because it's cleaner. We need some other way to get the functionality we want.